rename package from com.davideicardi.kaa to kaa.schemaregistry
