--- kaa/src/main/scala-2.12/com/davideicardi/kaa/utils/CollectionConverters.scala
@@ -1,9 +1,0 @@
-package com.davideicardi.kaa.utils
-
-import collection.JavaConverters._
-
-object CollectionConverters {
-    def mapAsJava[K, V](map: Map[K, V]) = {
-        map.asJava
-    }
-}
