+++ kaa/src/main/scala-2.12/kaa/schemaregistry/utils/CollectionConverters.scala
@@ -1,0 +1,9 @@
+package kaa.schemaregistry.utils
+
+import collection.JavaConverters._
+
+object CollectionConverters {
+    def mapAsJava[K, V](map: Map[K, V]) = {
+        map.asJava
+    }
+}
