--- kaa/src/main/scala-2.13/com/davideicardi/kaa/utils/CollectionConverters.scala
@@ -1,9 +1,0 @@
-package com.davideicardi.kaa.utils
-
-import scala.jdk.CollectionConverters._
-
-object CollectionConverters {
-    def mapAsJava[K, V](map: Map[K, V]) = {
-        map.asJava
-    }
-}
