+++ kaa/src/main/scala-2.13/kaa/schemaregistry/kaa/utils/CollectionConverters.scala
@@ -1,0 +1,9 @@
+package kaa.schemaregistry.utils
+
+import scala.jdk.CollectionConverters._
+
+object CollectionConverters {
+    def mapAsJava[K, V](map: Map[K, V]) = {
+        map.asJava
+    }
+}
