--- kaa/src/main/scala/com/davideicardi/kaa/SchemaId.scala
@@ -1,4 +1,0 @@
-package com.davideicardi.kaa
-
-case class SchemaId(value: Long) extends AnyVal {
-}
