--- kaa/src/main/scala/com/davideicardi/kaa/SchemaRegistry.scala
@@ -1,9 +1,0 @@
-package com.davideicardi.kaa
-
-import org.apache.avro.Schema
-
-trait SchemaRegistry {
-  def put(schema: Schema): SchemaId
-
-  def get(id: SchemaId): Option[Schema]
-}
