--- kaa/src/main/scala/com/davideicardi/kaa/exceptions.scala
@@ -1,6 +1,0 @@
-package com.davideicardi.kaa
-
-
-class InvalidSchemaException(s:String) extends Exception(s){}
-
-class SchemaNotFoundException(s:String) extends Exception(s){}
