+++ kaa/src/main/scala/kaa/schemaregistry/SchemaId.scala
@@ -1,0 +1,4 @@
+package kaa.schemaregistry
+
+case class SchemaId(value: Long) extends AnyVal {
+}
