+++ kaa/src/main/scala/kaa/schemaregistry/SchemaRegistry.scala
@@ -1,0 +1,9 @@
+package kaa.schemaregistry
+
+import org.apache.avro.Schema
+
+trait SchemaRegistry {
+  def put(schema: Schema): SchemaId
+
+  def get(id: SchemaId): Option[Schema]
+}
