+++ kaa/src/main/scala/kaa/schemaregistry/exceptions.scala
@@ -1,0 +1,6 @@
+package kaa.schemaregistry
+
+
+class InvalidSchemaException(s:String) extends Exception(s){}
+
+class SchemaNotFoundException(s:String) extends Exception(s){}
